Extract printMappingItem
